requirements.txt
boto3==1.9.141
botocore==1.12.141
Click==7.0
docutils==0.14
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.15.2
